Refactor SightSearchScreen
